requirements.txt
pygame
PyOpenGL
Pillow
pytest
cloc
